simulator/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.3.0
idna==3.7
numpy==1.26.4
requests==2.31.0
spring-config-client==0.2
urllib3==2.2.1
